// FE/src/shared/api/schema.d.ts
export interface components {
    schemas: {
// ...
        PlanOption: {
            type?: string;
            label?: string;
            description?: string;
            summaryReason?: string;
            days?: components["schemas"]["DayPlan"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}